requirements.txt
wget
openai-whisper